Created ga main as specified in the assignment doc. It takes in the arguments from the command line as stated in doc too. If you need to set arguments then go to the down arrow near the play button and hit run configurations, then select the tab arguments and set the first one to the puzzle you want to use (1, 2, or 3) then the second argument is the file name (in quotes if there is a space), and the third argument is the number in seconds the algorithm has to run.
The puzzleOne class has empty functions and a functional loop for stopping the algorithm when time runs out. Added a test.txt file as the text file for puzzleOne. Set up FileInputOutput.fileToPuzzleOne to take a file name and puzzleOne object and instantiate it with the info from the file.