NumberSequence: Changed fitness function to allow sums that are just above the goal to be fitter than a sum that is way over the goal.
PuzzleOne:
Added generation attribute to keep track of the number of generations the algorithm goes through.
Added a randomGenerator to the attributes to use over the whole class.
Added code to initialize the population, cull 10% of the weakest part of the population, and a psuedo reproduce function that will be switched out with spliciing later on.
Added a function to get the most fit object in the population,